rebase onto master & move tests to test_window.py
